+++ code/lib/cli/src/automigrate/helpers/eslintPlugin.ts
@@ -1,12 +1,12 @@
 import fse, { readFile, readJson, writeJson } from 'fs-extra';
-
 import { dedent } from 'ts-dedent';
 import detectIndent from 'detect-indent';
-import { readConfig, writeConfig } from '@storybook/csf-tools';
 import prompts from 'prompts';
 import chalk from 'chalk';
+
+import { readConfig, writeConfig } from '@storybook/csf-tools';
 import type { JsPackageManager } from '@storybook/core-common';
-import { paddedLog } from '../../helpers';
+import { paddedLog } from '@storybook/core-common';
 
 export const SUPPORTED_ESLINT_EXTENSIONS = ['js', 'cjs', 'json'];
 const UNSUPPORTED_ESLINT_EXTENSIONS = ['yaml', 'yml'];
