+++ code/lib/cli/src/automigrate/helpers/getMigrationSummary.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, vi, it } from 'vitest';
 import { getMigrationSummary } from './getMigrationSummary';
 import { FixStatus } from '../types';
-import type { InstallationMetadata } from '../../js-package-manager/types';
+import type { InstallationMetadata } from '@storybook/core-common';
 
 vi.mock('boxen', () => ({
   default: vi.fn((str, { title = '' }) => `${title}\n\n${str.replace(/\x1b\[[0-9;]*[mG]/g, '')}`),
