+++ code/lib/cli/src/automigrate/helpers/getMigrationSummary.ts
@@ -1,9 +1,9 @@
 import chalk from 'chalk';
 import boxen from 'boxen';
 import dedent from 'ts-dedent';
+import type { InstallationMetadata } from '@storybook/core-common';
 import type { FixSummary } from '../types';
 import { FixStatus } from '../types';
-import type { InstallationMetadata } from '../../js-package-manager/types';
 import { getDuplicatedDepsWarnings } from '../../doctor/getDuplicatedDepsWarnings';
 
 export const messageDivider = '\n\n';
