+++ code/lib/cli/src/automigrate/helpers/mainConfigFile.ts
@@ -11,7 +11,8 @@ import { readConfig, writeConfig as writeConfigFile } from '@storybook/csf-tools
 import chalk from 'chalk';
 import dedent from 'ts-dedent';
 import path from 'path';
-import type { getStorybookVersion, JsPackageManager } from '@storybook/core-common';
+import type { JsPackageManager } from '@storybook/core-common';
+import { getStorybookVersion } from '@storybook/core-common';
 
 const logger = console;
