+++ code/lib/cli/src/generators/types.ts
@@ -1,6 +1,6 @@
 import type { NpmOptions } from '../NpmOptions';
 import type { SupportedLanguage, Builder, ProjectType } from '../project_types';
-import type { JsPackageManager, PackageManagerName } from '../js-package-manager/JsPackageManager';
+import type { JsPackageManager, PackageManagerName } from '@storybook/core-common';
 import type { FrameworkPreviewParts } from './configure';
 
 export type GeneratorOptions = {
