+++ code/lib/cli/src/initiate.ts
@@ -4,6 +4,15 @@ import prompts from 'prompts';
 import { telemetry } from '@storybook/telemetry';
 import { withTelemetry } from '@storybook/core-server';
 import { NxProjectDetectedError } from '@storybook/core-events/server-errors';
+import {
+  versions,
+  HandledError,
+  JsPackageManagerFactory,
+  commandLog,
+  codeLog,
+  paddedLog,
+} from '@storybook/core-common';
+import type { JsPackageManager } from '@storybook/core-common';
 
 import dedent from 'ts-dedent';
 import boxen from 'boxen';
@@ -11,7 +20,6 @@ import { lt, prerelease } from 'semver';
 import type { Builder } from './project_types';
 import { installableProjectTypes, ProjectType } from './project_types';
 import { detect, isStorybookInstantiated, detectLanguage, detectPnp } from './detect';
-import { commandLog, codeLog, paddedLog } from './helpers';
 import angularGenerator from './generators/ANGULAR';
 import emberGenerator from './generators/EMBER';
 import reactGenerator from './generators/REACT';
@@ -28,13 +36,9 @@ import qwikGenerator from './generators/QWIK';
 import svelteKitGenerator from './generators/SVELTEKIT';
 import solidGenerator from './generators/SOLID';
 import serverGenerator from './generators/SERVER';
-import type { JsPackageManager } from '@storybook/core-common';
-import { JsPackageManagerFactory } from '@storybook/core-common';
 import type { NpmOptions } from './NpmOptions';
 import type { CommandOptions, GeneratorOptions } from './generators/types';
-import { HandledError } from './HandledError';
 import { currentDirectoryIsEmpty, scaffoldNewProject } from './scaffold-new-project';
-import { versions } from '@storybook/core-common';
 
 const logger = console;
