+++ code/lib/core-common/src/index.ts
@@ -14,6 +14,7 @@ export * from './utils/get-storybook-configuration';
 export * from './utils/get-storybook-info';
 export * from './utils/get-storybook-refs';
 export * from './utils/glob-to-regexp';
+export * from './utils/HandledError';
 export * from './utils/handlebars';
 export * from './utils/interpolate';
 export * from './utils/interpret-files';
@@ -22,6 +23,7 @@ export * from './utils/load-custom-presets';
 export * from './utils/load-main-config';
 export * from './utils/load-manager-or-addons-file';
 export * from './utils/load-preview-or-config-file';
+export * from './utils/log';
 export * from './utils/log-config';
 export * from './utils/normalize-stories';
 export * from './utils/paths';
