+++ code/lib/core-common/src/js-package-manager/JsPackageManager.ts
@@ -8,11 +8,11 @@ import fs from 'fs';
 import dedent from 'ts-dedent';
 import { readFile, writeFile, readFileSync } from 'fs-extra';
 import invariant from 'tiny-invariant';
-import { commandLog } from '../../../cli/src/helpers';
+import { commandLog } from '../utils/log';
 import type { PackageJson, PackageJsonWithDepsAndDevDeps } from './PackageJson';
 import storybookPackagesVersions from '../versions';
 import type { InstallationMetadata } from './types';
-import { HandledError } from '../../../cli/src/HandledError';
+import { HandledError } from '../utils/HandledError';
 
 const logger = console;
