+++ code/lib/core-common/src/utils/HandledError.ts
